H5Easy API (#141)
This commit adds the initial functionality around the concept of a high level API for HighFive. Typical cases are import, export and format conversion.

For the specific case of import and export two main free functions are defined: `load` and `dump`, which support the following formats:
 * Plain scalar & std::vectors
 * Xtensor arrays
 * Eigen Matrices

After importing `highfive/H5Easy.hpp` functions are available in the new `H5Easy` namespace.

Further, this PR adds a couple of handy shortcuts in the core `H5DataSet.hpp` namely `getDimensions()` and `getElementCount()` previously available only from a DataSpace object.

Thanks everyone for the long contribution process and reviews.