Fix another MSVC warnings when building for win32 (#179)
After PR #169 there is still a warning remaining when building for win32
where sizeof(size_t) > sizeof(unsigned int). Fix it by using brace
initialization which requires that the value inside is exactly
representable in size_t. This is better than the static_cast used in the
initial version of PR #169.