Fix MSVC issue with 2D Vectors (#193)
Change signature of the functions to force a different name mangling which allows MSVC to compile the code. The issue, according to what I've been able to find, is that MSVC cannot disambiguate between a const vs non const function parameter when doing name mangling. Clang and GCC use a different mangling and are able to tell the difference and as such can disambiguate between the candidates. GCC and Clang have conforming behavior and MSVC does not.

Fixes #171 and #104

I reported the issue to Microsoft here https://developercommunity.visualstudio.com/content/problem/736092/standards-conforming-code-returns-c2668-ambiguous.html
It is highly unlikely that MS will be able to fix this issue as it is core to how their name mangling works and likely would cause ABI issues to make changes.