Added method `getFile()` to every object (except `DataType` yet)
If we have something derived from `object` then sometimes it is convenient to get file that this object is belong to. So I added such method `getFile()`.
I addede if to each of the object and to the base class `object` as class `file` is also derived from object and that would be cumbersome.

The only thing I can't overcome now is to add this method to `datatype` class. I eed somehow to `include "h5file.hpp"` but I can't as I don't fully understand how the library is structured.

Also I propose to remove private member `_filename` as we always can retrieve it via `H5Fget_name(...)` method.