Added factory object's constructor FromId(), added == and != operators
To overcome include recursion that was before I decided to add constructor to each object-derived that works likeObject::FromId(id). But I had to add protected constructor like Object(object) because of Dataspace(). For example without that protected constructor trick initializing Dataspace(1) would try to invoke constructor from ID.

I also added == and != operators for object and Dataspace. Object operators tries to compare adress (wich is deprecated) of object within HDF5 file and then compare if these objects reside in the same HDF5 or not.
For Dataspace it checks the dimensions equality.

I removed private File's var _filename because it increase the size of FIle (from 8 to 56 bytes as I remember) as it is always possible retrieve this information via standard HDF5 functionality.

Added tests `HighFiveFilename` and `HighFiveCheckIdFunctionality`