Add GHA workflow to check doxygen-awesome version
Rename `download_stylesheet.sh` to `update_doxygen_awesome.sh`. Since
the script must check out the doxygen-awesome repo, it need to be passed
a folder that is 'safe' to pollute.

The workflow will download the most current version of the CSS. Then
check if it's different from our version. It will proceed to create a
pull request, but only if it hasn't done so on a previous run.