Fall back to a Catch2 in the CMake paths
@tdegeus, I think this should fix #527 sufficiently.  Any CI setup
therein should maybe be followed up separately.  A bit of a butchered
CMake in my opinion, but given the "language" constraints, it seems to
work.

Will set unit tests to AUTO, disabling them if neither the submodule nor
a system package for Catch2 is found.  The submodule will be preferred,
otherwise.

Errors out if unit tests are set to ON and no Catch2 is found.