Implement logging infrastructure. (#690)
Implements basic logging infrastructure for human readable amounts of messages. The design idea is that any user of HighFive can optionally specify a backend logger by registering a callback.

The logging API consists of:

    HIGHFIVE_LOG_{DEBUG,INFO,WARN,ERROR}(message)
    HIGHFIVE_LOG_*_IF(cond, message)

The former forwards the message with the appropriate log severity along with the filename and line number to the logging callback. The latter does so only if the `if`-statement is true. It's important to avoid any side effects in `cond` since the macro might expand to a no-op if the compile time minimum log severity is too high. Therefore, `cond` wont be evaluated and the side effects wont happen.