Fixes broadcasting of arrays. (#697)
Fixes broadcasting of arrays.

The rules for broadcasting in `2.4.x` seem to be that singleton
dimension can be stripped from either side. If both is possible
it'll strip from the back, i.e. the dataset with shape `(1, 3, 1)`
can be read into an array of shape `(1, 3)`. This was probably
motivated by `std::vector<std::vector<double>>` of shape `(1, 3)`
is easier to deal with than one that's `(3, 1)`.

However, the stripping from both sides only works for one-dimensional
arrays. Stripping multi-dimensional arrays from either side is unstable
and should be avoided anyway. Hence, we only allow stripping any `1`s
if the target shape is one-dimensional.

This commit achieves the above by:
  * introducing `details::squeezeDimensions` which strips the
    required amount of singleton dimesions.

  * improving `details::checkDimensions` by removing bugs related
    to `(3, 1, 1)` not being a valid 2D array.

  * Adds (back) tests for checking the broadcasting rules. As well as
    `details::checkDimensions` and `details::squeezeDimensions`.