Remove Eigen functionality for matrices.
HighFive hasn't been respecting that Eigen uses Fortran order, as a
result it's been writing `Eigen::Matrix` to file as if the data were C
order. It's been consistently reading datasets back into `Eigen::Matrix`
has if it already was in Fortran order.

This error is symmetric, i.e., if one writes an `Eigen::Matrix` to disk
via core HighFive and then reads that dataset back into an
`Eigen::Matrix` the outcome is correct (which has been check in CI).

More information can be found at:
https://github.com/BlueBrain/HighFive/issues/532