Make codecov CI more permissive. (#736)
Code coverage in HighFive should not determine the success or failure of
CI. It might be useful to notice if CI essentially doesn't cover
anything anymore. Hence the threshold of 50%.

The setting 'fail_ci_if_error' will make the CI fail if there are issue
uploading results. Hence, this will just make the CI fail spuriously,
e.g. if codecov servers are down. The default is false, we should stick
with it.

The 'target' is set to zero, we leave it to human reviewers to determine
a satisfactory level of coverage.