Refactor raw read and write code.
A raw read should be a very thin wrapper around HDF5. This allows users
to use HighFive (as opposed to pure HDF5) to simply write from or read
into a pointer.

To enable this, the function are restructure to not use the inspector if
the user specified all required information.