Improve CMake `HIGHFIVE_HAS_CONCEPTS`.
The commit sets the variable for the target `libdeps` (and transitively
`HighFive`). Furthermore, it ensures that the value of
`HIGHFIVE_HAS_CONCEPTS` is either `0` or `1`. Finally, it removes the
check for C++ standard, because we already check if the compiler should
be able to handle C++20 features via `__cplusplus` in the code.