Distinguish between buffer pointer and iterator. (#805)
There's a notion of pointer to the first byte of the buffer. This
is what one passes to HDF5, e.g. in `H5Dread`. There's also the concept
of a iterator which is used for serialization.

Currently, both ideas are represented by an `hdf5_type*`. However, the
StringBuffer will need to return proxy objects. This commits splits the
notion. Conveniently, we can rename `get_pointer` to the more in-style
`getPointer` and hence the compiler checks that we've considered all
occurrences of `get_pointer`.