Update Github templates.
Bug report: Slightly changes the wording and what versions we need.
Feature request: This can be free form text.
Pull request: Deleted in favour of simply having the commit title and
description.
Build failure: New, contains the information we need for build failure
and instructions on how to create a reproducer.