Fail when reference counts are inconsistent.
In HighFive we have two types of invalid HIDs. The first category
consists of:
* `H5I_INVALID_HID` which HighFive uses to indicate that the object is
  "empty",
* `H5P_DEFAULT` (aka `0`) which is to "optimize" the default property.

The second category consists of anything else that `H5Iis_valid` reports
as invalid, e.g. identifiers that have been closed, etc.

For the first category we must not increment or decrement the reference
counters. However, for the second category, we'd like to get an error,
because we're expecting either `H5I_INVALID_HID`, `H5P_DEFAULT` or a
valid HID, but are given something else.