Remove `inspector<T>::getSize{,Val}`. (#959)
The method getSizeVal(T& value) is problematic because some containers don't know their dimensions, yet they need to forward the input pointer when serializing/deserializing. After changing serialize to accept the dims the number of elements can be obtained from there. It's also tricky in the context of arrays of chars, these could be a counted as one string or as multiple characters. Inside the inspector we don't know which interpretation is correct.

The method getSize(dims) seems to only have a single plausible implementation. Hence, it's being removed mostly for cosmetic reasons.