Windows CI should be red on test failure.
The issue is that `bash -l {0}` seems to turn off the `set -e` behaviour
we see in the other pipelines. Simply asking for `bash`, leads to every
ctest invocation failing with an exit code (even tests that are know to
succeed).

Therefore, we put one test per stage of the pipeline, this ensures its
exit code is relevant for success/failure of the pipeline.